Maybe transfer key, values on incoming request.
The existing code only checks if we should call transferKeyValues when we
receive a response to one of our RPCs. We also need to check when receive a
new request from a new node.

This commit adds a new method that is called each time we receive a new request.